refactors typescript components and import paths
