+++ demo/components/simple-heading/simple-heading.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import Link from 'components/link';
+import Link from 'components/link/link';
 import './simple-heading.scss';
 
 /**
