+++ demo/views/chrome/header/header.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Button from 'components/button';
+import Button from 'components/button/button';
 import Icon from 'components/icon';
 
 import DemoActions from '../../../actions/demo';
