+++ demo/views/pages/home/sage-loves-carbon/sage-loves-carbon.js
@@ -3,7 +3,7 @@ import devices from './devices.png';
 import React from 'react';
 import I18n from 'i18n-js';
 
-import Link from 'components/link';
+import Link from 'components/link/link';
 
 import ComplexHeading from './../../../../components/complex-heading';
 import Wrapper from './../../../common/wrapper';
