+++ src/components/action-toolbar/action-toolbar.js
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
 import I18n from 'i18n-js';
-import Link from '../link';
+import Link from '../link/link';
 import tagComponent from '../../utils/helpers/tags';
 import './action-toolbar.scss';
