+++ src/components/configurable-items/configurable-items.js
@@ -4,7 +4,7 @@ import classNames from 'classnames';
 import I18n from 'i18n-js';
 import tagComponent from '../../utils/helpers/tags';
 import { DraggableContext } from '../drag-and-drop';
-import Button from '../button';
+import Button from '../button/button';
 import ConfigurableItemRow from './configurable-item-row';
 import Form from '../form';
 import './configurable-items.scss';
