+++ src/components/create/create.js
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
 import tagComponent from '../../utils/helpers/tags';
-import Link from '../link';
+import Link from '../link/link';
 import './create.scss';
 
 class Create extends React.Component {
