+++ src/components/link/__definition__.js
@@ -1,4 +1,4 @@
-import Link from './';
+import Link from './link';
 import Definition from './../../../demo/utils/definition';
 import OptionsHelper from '../../utils/helpers/options-helper';
 import linkDefinition from './../link/__definition__';
