+++ src/components/link/index.ts
@@ -1,0 +1,1 @@
+export { default as Link } from './link';
