+++ src/components/menu/menu-item/menu-item.js
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
 import { assign } from 'lodash';
-import Link from '../../link';
+import Link from '../../link/link';
 import tagComponent from '../../../utils/helpers/tags';
 import './menu-item.scss';
