+++ src/components/multi-action-button/multi-action-button.js
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
 import Icon from '../icon';
-import Button from '../button';
+import Button from '../button/button';
 import SplitButton from '../split-button';
 import './multi-action-button.scss';
